#11 refactor : Week리스트 구현 로직 수정
