+++ src/components/dashboard/hooks/useMediaLoad.ts
@@ -1,6 +1,7 @@
 import React from 'react';
 import { getTotalDataAPI } from 'libs/api/getDataAPI';
 import { MediaData } from 'types/dashboard';
+import { INITIAL_WEEK_STATE } from 'libs/utils/constants';
 
 const useMediaLoad = (gteDate: string, lteDate: string) => {
   // locaohost:json-server/report의 모든 데이터를 저장하는 state: Week 데이터로 가공하기 전 상태
@@ -10,19 +11,19 @@ const useMediaLoad = (gteDate: string, lteDate: string) => {
   React.useEffect(() => {
     const getMediaData = async () => {
       try {
-        const totalData = await getTotalDataAPI<MediaData[]>(
+        const mediaResponse = await getTotalDataAPI(
           'media',
           `?date_gte=${gteDate}&date_lte=${lteDate}`,
         );
-        setTotalData(totalData);
+        setTotalData(mediaResponse.data);
       } catch (error) {
         console.log(error);
       }
     };
-    if (gteDate && lteDate) {
+    if (lteDate !== INITIAL_WEEK_STATE[1]) {
       getMediaData();
     }
-  }, [gteDate, lteDate]);
+  }, [lteDate]);
 
   return { totalDataContainingDates };
 };
