+++ src/libs/api/getDataAPI.ts
@@ -1,10 +1,7 @@
 import apiClient from '.';
 
 // report 데이터를 받아오는 api
-export const getTotalDataAPI = async <T>(
-  property: string,
-  param = '',
-): Promise<T> => {
+export const getTotalDataAPI = async (property: string, param = '') => {
   const response = await apiClient.get(`/${property}${param}`);
-  return response.data;
+  return response;
 };
